Infra/requirements.txt
python-dotenv
requests
azure-storage-blob
